requirements.txt
transformers
langchain
langchain_community
langchain-openai
torch
pypdf
sentence-transformers
faiss-cpu
openai
python-dotenv
pytesseract
unstructured
SpeechRecognition
fastapi
uvicorn[standard]
streamlit
streamlit-chat
python-docx
requests
tqdm